+++ crates/full-node/sov-metrics/src/lib.rs
@@ -1,4 +1,4 @@
-//! Metric tracking for Sovereign rollups.
+#![doc = include_str!("../README.md")]
 #![deny(missing_docs)]
 
 /// Contains utilities to track zkvm cycles
@@ -10,15 +10,15 @@ mod influxdb;
 #[cfg(feature = "native")]
 pub use influxdb::{
     init_metrics_tracker, timestamp, track_metrics, BatchMetrics, BatchOutcome, HttpMetrics,
-    MetricsTracker, MonitoringConfig, RunnerMetrics, SlotProcessingMetrics, SovRollupMetrics,
-    TransactionEffect, TransactionProcessingMetrics, UserSpaceSlotProcessingMetrics, ZkCircuit,
-    ZkProvingTime, ZkVmExecutionChunk, METRICS_METADATA,
+    Metric, MetricsTracker, MonitoringConfig, RunnerMetrics, SlotProcessingMetrics,
+    SovRollupMetrics, TransactionEffect, TransactionProcessingMetrics,
+    UserSpaceSlotProcessingMetrics, ZkCircuit, ZkProvingTime, ZkVmExecutionChunk, METRICS_METADATA,
 };
 #[cfg(all(feature = "native", feature = "gas-constant-estimation"))]
 pub use influxdb::{GasConstantTracker, GAS_CONSTANTS};
 
 #[macro_export]
-/// Starts a timer if the `native` feature is enabled. Otherwise does nothing.
+/// Starts a timer if the `native` feature is enabled. Otherwise, does nothing.
 macro_rules! start_timer {
     ($timer:ident) => {
         #[cfg(feature = "native")]