+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #2522 includes git hash as a resource attribute in metrics so we know exactly what version of the rollup is running.
 - #2533 Changes demo-stf dependency usage from path to workspace.
 - #2537 Adjusts amounts in the testing framework to use `Amount` instead of u128
+- #2539 Adds support for custom metrics tracking. Please refer to README in `sov-metrics` crate.
 
 ## 2025-02-26
 - #2423 integrates the `AccessPattern` module into the demo-stf `Runtime`.